Refactoring code and Adding permission validations
